--- pypeit/bspline/__init__.py
@@ -1,3 +1,0 @@
-
-from pypeit.bspline.bspline import bspline
-
