+++ pypeit/core/bspline/__init__.py
@@ -1,0 +1,3 @@
+
+from pypeit.core.bspline.bspline import bspline
+
