+++ pypeit/core/skysub.py
@@ -10,12 +10,12 @@ import scipy.ndimage
 import scipy.special
 from scipy.interpolate import RegularGridInterpolator
 
-from pypeit import bspline
 from pypeit import log
 from pypeit import PypeItError
 from pypeit import slittrace
 from pypeit import utils
 from pypeit.core import basis
+from pypeit.core import bspline
 from pypeit.core import extract
 from pypeit.core import fitting
 from pypeit.core import pixels
