+++ pypeit/flatfield.py
@@ -22,7 +22,7 @@ from IPython import embed
 from pypeit import log
 from pypeit import PypeItError, PypeItDataModelError
 from pypeit import utils
-from pypeit import bspline
+
 
 from pypeit import datamodel
 from pypeit import calibframe
@@ -30,6 +30,7 @@ from pypeit import edgetrace
 from pypeit import io
 from pypeit.display import display
 from pypeit.images import buildimage
+from pypeit.core import bspline
 from pypeit.core import qa
 from pypeit.core import flat
 from pypeit.core import tracewave
@@ -71,7 +72,7 @@ class FlatImages(calibframe.CalibFrame):
                  'pixelflat_model': dict(otype=np.ndarray, atype=np.floating, descr='Model flat'),
                  'pixelflat_spat_bsplines': dict(otype=np.ndarray, atype=bspline.bspline,
                                                  descr='B-spline models for pixel flat; see '
-                                                       ':class:`~pypeit.bspline.bspline.bspline`'),
+                                                       ':class:`~pypeit.core.bspline.bspline.bspline`'),
                  'pixelflat_finecorr': dict(otype=np.ndarray, atype=fitting.PypeItFit,
                                        descr='PypeIt 2D polynomial fits to the fine correction of '
                                              'the spatial illumination profile'),
@@ -85,7 +86,7 @@ class FlatImages(calibframe.CalibFrame):
                                        descr='Processed, combined illum flats'),
                  'illumflat_spat_bsplines': dict(otype=np.ndarray, atype=bspline.bspline,
                                                  descr='B-spline models for illum flat; see '
-                                                       ':class:`~pypeit.bspline.bspline.bspline`'),
+                                                       ':class:`~pypeit.core.bspline.bspline.bspline`'),
                  'illumflat_finecorr': dict(otype=np.ndarray, atype=fitting.PypeItFit,
                                        descr='PypeIt 2D polynomial fits to the fine correction of '
                                              'the spatial illumination profile'),
